completed, waiting for testing

+++ cogs/match.py
@@ -229,12 +229,18 @@ class ReadyButton(ui.View):
             # CHECK
             if len(ready_ups) == 2:
             # if len(ready_ups) == 10:
+                preference = await self.bot.fetchrow(f"SELECT * FROM queue_preference WHERE guild_id = {inter.guild.id}")
+                if preference:
+                    preference = preference[1]
+                else:
+                    preference = 1
 
-                for member in game_members:
-                    # Remove member from all other queues
-                    await self.bot.execute(
-                        f"DELETE FROM game_member_data WHERE author_id = {member} and game_id != '{self.game_id}'"
-                    )
+                if preference == 1:
+                    for member in game_members:
+                        # Remove member from all other queues
+                        await self.bot.execute(
+                            f"DELETE FROM game_member_data WHERE author_id = {member} and game_id != '{self.game_id}'"
+                        )
 
                 await self.bot.execute(
                     f"DELETE FROM ready_ups WHERE game_id = '{self.game_id}'"
@@ -342,6 +348,7 @@ class ReadyButton(ui.View):
                 )
 
                 self.disable_button.cancel()
+                await Match.start(self, inter.channel)
 
         else:
             await inter.send(
@@ -393,7 +400,33 @@ class QueueButtons(ui.View):
             return True
         return False
 
+    async def in_ongoing_game(self, inter) -> bool:
+        data = await self.bot.fetch(f"SELECT * FROM games")
+        for entry in data:
+            user_roles = [x.id for x in inter.author.roles]
+            if entry[4] in user_roles or entry[5] in user_roles:
+                return True
+
+        return False
+
     async def add_participant(self, inter, button) -> None:
+        preference = await self.bot.fetchrow(f"SELECT * FROM queue_preference WHERE guild_id = {inter.guild.id}")
+        if preference:
+            preference = preference[1]
+        else:
+            preference = 1
+        
+        if preference == 2:
+            in_other_games = await self.bot.fetch(
+                f"SELECT * FROM game_member_data WHERE author_id = {inter.author.id} and game_id != '{self.game_id}'"
+            )
+            if in_other_games:
+                return await inter.send(
+                    embed=error(f"You cannot be a part of multiple queues."),
+                    ephemeral=True,
+                )
+
+
         label = button.label.lower()
         team = "blue"
 
@@ -439,7 +472,7 @@ class QueueButtons(ui.View):
 
         # CHECK
         if checks_passed == 1:
-        # if checks_passed == len(self.children) - 2:
+        # if checks_passed == len(self.children) - 1:
             member_data = await self.bot.fetch(
                 f"SELECT * FROM game_member_data WHERE game_id = '{self.game_id}'"
             )
@@ -470,11 +503,12 @@ class QueueButtons(ui.View):
                 else:
                     rating = Rating()
                     await self.bot.execute(
-                        f"INSERT INTO mmr_rating(guild_id, user_id, mu, sigma) VALUES($1, $2, $3, $4)",
+                        f"INSERT INTO mmr_rating(guild_id, user_id, mu, sigma, counter) VALUES($1, $2, $3, $4, $5)",
                         inter.guild.id,
                         data[0],
                         rating.mu,
-                        rating.sigma
+                        rating.sigma,
+                        0
                     )
 
                 roles_occupation[data[1].upper()].append({'user_id': data[0], 'rating': rating})
@@ -535,6 +569,9 @@ class QueueButtons(ui.View):
         await inter.response.defer()
         if not self.game_id:
             self.game_id = inter.message.embeds[0].footer.text
+        
+        if await self.in_ongoing_game(inter):
+            return await inter.send(embed=error("You are already in an ongoing game."), ephemeral=True)
 
         game_members = await self.bot.fetch(
             f"SELECT * FROM game_member_data WHERE game_id = '{self.game_id}'"
@@ -619,40 +656,40 @@ class QueueButtons(ui.View):
                 embed=error("You are not a participant of this game."), ephemeral=True
             )
 
-    @ui.button(label="Switch Team", style=ButtonStyle.blurple, custom_id="switchteam")
-    async def switchteam(self, button, inter):
-        await inter.response.defer()
-        data = await self.bot.fetchrow(
-            f"SELECT * FROM game_member_data WHERE author_id = {inter.author.id} and game_id = '{self.game_id}'"
-        )
-        if data:
+    # @ui.button(label="Switch Team", style=ButtonStyle.blurple, custom_id="switchteam")
+    # async def switchteam(self, button, inter):
+    #     await inter.response.defer()
+    #     data = await self.bot.fetchrow(
+    #         f"SELECT * FROM game_member_data WHERE author_id = {inter.author.id} and game_id = '{self.game_id}'"
+    #     )
+    #     if data:
 
-            check = await self.bot.fetchrow(
-                f"SELECT * FROM game_member_data WHERE role = '{data[1]}' and game_id = '{self.game_id}' and author_id != {inter.author.id}"
-            )
-            if check:
-                return await inter.send(
-                    "The other team position for this role is already occupied.",
-                    ephemeral=True,
-                )
+    #         check = await self.bot.fetchrow(
+    #             f"SELECT * FROM game_member_data WHERE role = '{data[1]}' and game_id = '{self.game_id}' and author_id != {inter.author.id}"
+    #         )
+    #         if check:
+    #             return await inter.send(
+    #                 "The other team position for this role is already occupied.",
+    #                 ephemeral=True,
+    #             )
 
-            if data[2] == "blue":
-                team = "red"
-            else:
-                team = "blue"
+    #         if data[2] == "blue":
+    #             team = "red"
+    #         else:
+    #             team = "blue"
 
-            await self.bot.execute(
-                f"UPDATE game_member_data SET team = '{team}' WHERE game_id = $1 and author_id = $2",
-                self.game_id,
-                inter.author.id,
-            )
-            await inter.edit_original_message(embed=await self.gen_embed(inter.message))
-            await inter.send(f"You were assigned to **{team} team**.", ephemeral=True)
+    #         await self.bot.execute(
+    #             f"UPDATE game_member_data SET team = '{team}' WHERE game_id = $1 and author_id = $2",
+    #             self.game_id,
+    #             inter.author.id,
+    #         )
+    #         await inter.edit_original_message(embed=await self.gen_embed(inter.message))
+    #         await inter.send(f"You were assigned to **{team} team**.", ephemeral=True)
 
-        else:
-            await inter.send(
-                embed=error("You are not a part of this game."), ephemeral=True
-            )
+    #     else:
+    #         await inter.send(
+    #             embed=error("You are not a part of this game."), ephemeral=True
+    #         )
 
 
 class Match(Cog):
@@ -662,33 +699,33 @@ class Match(Cog):
 
     def __init__(self, bot):
         self.bot = bot
-        self.start_queue.start()
+        # self.start_queue.start()
 
-    @tasks.loop(seconds=5)
-    async def start_queue(self):
-        await self.bot.wait_until_ready()
+    # @tasks.loop(seconds=5)
+    # async def start_queue(self):
+    #     await self.bot.wait_until_ready()
 
-        try:
-            # May start before bot is ready, if tables are not ready, this may cause error
-            channels = await self.bot.fetch("SELECT * FROM queuechannels")
-        except:
-            channels = []
+    #     try:
+    #         # May start before bot is ready, if tables are not ready, this may cause error
+    #         channels = await self.bot.fetch("SELECT * FROM queuechannels")
+    #     except:
+    #         channels = []
 
-        for channel in channels:
-            channel = self.bot.get_channel(channel[0])
-            if not channel:
-                continue
-            data = await self.bot.fetch(
-                f"SELECT game_id FROM games WHERE queuechannel_id = {channel.id}"
-            )
-            if not data:
-                continue
-            async for msg in channel.history(limit=200):
-                if msg.embeds:
-                    if msg.embeds[0].footer:
-                        if data[-1][0] == msg.embeds[0].footer.text:
-                            await self.start(channel)
-                        break
+    #     for channel in channels:
+    #         channel = self.bot.get_channel(channel[0])
+    #         if not channel:
+    #             continue
+    #         data = await self.bot.fetch(
+    #             f"SELECT game_id FROM games WHERE queuechannel_id = {channel.id}"
+    #         )
+    #         if not data:
+    #             continue
+    #         async for msg in channel.history(limit=200):
+    #             if msg.embeds:
+    #                 if msg.embeds[0].footer:
+    #                     if data[-1][0] == msg.embeds[0].footer.text:
+    #                         await self.start(channel)
+    #                     break
 
     @Cog.listener()
     async def on_ready(self):
